closed_loop_enzyme_bench/requirements.txt
transformers>=4.35
accelerate
torch
biopython
pandas
numpy
matplotlib
tqdm
scikit-learn
pyyaml
